First crack at new build system
The new layout has the DSFMLC source files included in this repository. It
also changes the way DSFMLC is built, no longer being built on top of
SFML's source. Instead, SFML is a submodule of this repository and is
require in order to build correctly.

build.d has started to be reworked to be more useful. It will build both
dsfml libs and dsfmlc libs. It also now detects when a single file has
been changed, and will rebuild that file before linking.