Second crack at new build system.
This updates the build script and cmake script to work on linux, and updates
the travis file to utilize the new system.

This also changes the ftp server that is used in the unittest since the
previous one went down, and updates the DSFMLC Texture.cpp file to make use of
Linux's OpenGL to make flush work.